+++ src/model/content-type.ts
@@ -1,4 +1,4 @@
-import { MediaType } from "../utils/media-type"
+import { getStructuredMediaTypePredicate, MediaType } from "../utils/media-type"
 
 export type WithContentType = {
   contentType: MediaType,
@@ -6,4 +6,8 @@ export type WithContentType = {
 
 export function createMediaType(contentTypeHeader: string): MediaType {
   return new MediaType(contentTypeHeader);
+}
+
+export function hasStructure<T extends WithContentType>(type: string, suffix: string): (obj: T) => boolean {
+  return (obj: T) => getStructuredMediaTypePredicate(type, suffix)(obj.contentType);
 }