+++ src/model/headers.ts
@@ -1,4 +1,4 @@
-import { condition, setProperty } from "../utils/functions";
+import { choose, condition, resolve, setProperty } from "../utils/functions";
 
 export type WithHeadersClass = {
   headers: Headers,
@@ -18,6 +18,10 @@ export function hasHeader<T extends WithHeaders>(key: string): (obj: T) => boole
 
 export function withHeader<T extends WithHeaders, U>(key: string, fn: (obj: T, value: string) => U): (obj: T) => T | U {
   return condition(hasHeader(key), (obj: T) => fn(obj, obj.headers[key]));
+}
+
+export function withHeaderAsync<T extends WithHeaders, U>(key: string, fn: (obj: T, value: string) => Promise<U>): (obj: T) => Promise<T | U> {
+  return choose(hasHeader(key), (obj: T) => fn(obj, obj.headers[key]), resolve);
 }
 
 export function withHeaderSet<T extends WithHeaders, K extends PropertyKey, V>(key: string, propertyKey: K, valueFn: (value: string) => V): (obj: T) => T | (T & { [key in K]: V }) {